# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "DigiSchool API",
        "description": "API REST pour la gestion d'ecoles DigiSchool.\n\n## Authentification\nLa plupart des endpoints necessitent un token JWT.\n1. Utilisez `POST /api/auth/login` pour obtenir un token\n2. Cliquez sur le bouton \"Authorize\" et entrez: `Bearer <votre_token>`\n\n## Multi-tenancy\nLe systeme est multi-tenant base sur l'ecole (ecoleId).\nL'ecoleId est automatiquement extrait du token JWT.\n\n## Comptes de test\n| Role | Email | Telephone | Password |\n|------|-------|-----------|----------|\n| Admin | admin@digischool.cm | +237600000000 | Admin@2025 |\n| Directeur | smbarga@lavictoire.cm | +237677123456 | Directeur@2025 |\n| Enseignant | jpkamga@lavictoire.cm | +237670111222 | Enseignant@2025 |\n| Secretaire | catangana@lavictoire.cm | +237660555666 | Secretaire@2025 |\n| Parent | fnkoulou@gmail.com | +237691666777 | Parent@2025 |\n",
        "contact": {
            "name": "DigiSchool Support",
            "url": "https://digischool.cm",
            "email": "support@digischool.cm"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://digischool.cm/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Serveur de developpement local"
        },
        {
            "url": "https://api.digischool.cm",
...
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
        "/api/villes/{id}": {
            "get": {
                "tags": [
                    "ville-controller"
                ],
                "operationId": "getById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VilleDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ville-controller"
                ],
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VilleDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VilleDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ville-controller"
                ],
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/villes": {
            "get": {
                "tags": [
                    "ville-controller"
                ],
                "operationId": "getAll",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/VilleDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ville-controller"
                ],
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VilleDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VilleDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/villes/arrondissement/{arrondissementId}": {
            "get": {
                "tags": [
                    "ville-controller"
                ],
                "operationId": "getByArrondissementId",
                "parameters": [
                    {
                        "name": "arrondissementId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/VilleDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/quartiers/ville/{villeId}": {
            "get": {
                "tags": [
                    "quartier-controller"
                ],
                "operationId": "getByVilleId",
                "parameters": [
                    {
                        "name": "villeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/QuartierDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "VilleDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "code": {
                        "type": "string"
                    },
                    "nom": {
                        "type": "string"
                    },
                    "codePostal": {
                        "type": "string"
                    },
                    "arrondissementId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "arrondissementNom": {
                        "type": "string"
                    }
                }
            },
...
            "QuartierDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "code": {
                        "type": "string"
                    },
                    "nom": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "villeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "villeNom": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT obtenu via POST /api/auth/login. Entrez: Bearer <token>",
                "name": "bearerAuth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}